tradefed: compute stats only at the end
Updating testrun status while there's still lots of tests going on
is very costly database-wise due to atomic transactions. This causes
other threads to wait until a lock of the object is released.

This was removed and a new call to RecordTestRunStatus is made at the very
end, so the status can be calculated in a single step.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>