tradefed: adjust to SQUAD's new way to handle subtasks
The tradefed plugin hasa mechanism that triggers subtasks in parallel to
work on CTS/VTS tests. When subtasks are created, the main task reaches
the end, giving SQUAD the impression that all work is done, when in reality
there is still a lot going on subtasks.

This patch introduces a call to SQUAD's `update_testjob_status` task from
within `tradefed.tasks.update_build_status` to signal SQUAD that a testjob
status can be updated.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>