tradefed: lower test batch to 100
The reason for this is scaled up deployments. On a full power scenario
SQUAD may have several workers, with multiple threads within. This means
that dozens of batch test inserts will happen at the same time, causing
higher latency in workers.

If this number comes to be changed again, it might be a good idea to parameterize it.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>